+++ src/main/java/carpet/api/settings/SettingsManager.java
@@ -52,7 +52,7 @@ import net.minecraft.commands.SharedSuggestionProvider;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.level.storage.LevelResource;
-import org.thinkingstudio.bedsheet.util.NeoHooks;
+import org.thinkingstudio.bedsheet.loader.FoxifiedLoader;
 
 import static carpet.utils.Translations.tr;
 import static java.util.Comparator.comparing;
@@ -461,9 +461,9 @@ public class SettingsManager {
         }
         catch (NoSuchFileException e)
         {
-            if (path.equals(getFile()) && NeoHooks.getEnvironmentType().isClient())
+            if (path.equals(getFile()) && FoxifiedLoader.getEnvironmentType().isClient())
             {
-                Path defaultsPath = NeoHooks.getConfigDir().resolve("carpet/default_"+identifier+".conf");
+                Path defaultsPath = FoxifiedLoader.getConfigDir().resolve("carpet/default_"+identifier+".conf");
                 try {
                     if (Files.notExists(defaultsPath))
                     {