+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
-- port to NeoForge 1.21.5
+- change: `NeoHooks` -> `FoxifiedLoader`
+- NeoForge version `21.5.8-beta` -> `21.5.39-beta`